requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# RAG and LLM
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10

# Vector database
faiss-cpu>=1.9.0
chromadb==0.4.18

# Document processing
pypdf==3.17.4
python-docx==1.1.0
unstructured==0.11.6

# Korean text processing (custom implementation)

# Embeddings
sentence-transformers==2.2.2
openai>=1.6.1,<2.0.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
numpy>=1.24.0
pandas>=2.0.0
tqdm==4.66.1

# Development
pytest==7.4.3
black==23.11.0
ruff==0.1.6